requirements.txt
pandas
numpy
matplotlib
seaborn
mysql-connector-python
jupyter
streamlit
plotly
geopy
streamlit-folium
folium